Creating Configuration, Adding services and test main.class

+++ spring-data/src/main/java/by/pvt/repository/BodyTypeRepository.java
@@ -4,5 +4,5 @@ package by.pvt.repository;
 import by.pvt.pojo.BodyType;
 import org.springframework.data.repository.CrudRepository;
 
-public interface BodyTypeRepository extends CrudRepository<BodyType, Long> {
+public interface BodyTypeRepository extends CrudRepository<BodyType, Integer> {
 }
