+++ spring-data/src/main/java/by/pvt/repository/CarRepository.java
@@ -3,6 +3,6 @@ package by.pvt.repository;
 import by.pvt.pojo.Car;
 import org.springframework.data.repository.CrudRepository;
 
-public interface CarRepository extends CrudRepository<Car,Long>{
+public interface CarRepository extends CrudRepository<Car, Integer>{
 	
 }
