+++ spring-data/src/main/java/by/pvt/repository/EngineTypeRepository.java
@@ -3,5 +3,5 @@ package by.pvt.repository;
 import by.pvt.pojo.EngineType;
 import org.springframework.data.repository.CrudRepository;
 
-public interface EngineTypeRepository extends CrudRepository<EngineType, Long> {
+public interface EngineTypeRepository extends CrudRepository<EngineType, Integer> {
 }
