+++ spring-data/src/main/java/by/pvt/repository/StatusRepository.java
@@ -3,5 +3,5 @@ package by.pvt.repository;
 import by.pvt.pojo.OrderStatus;
 import org.springframework.data.repository.CrudRepository;
 
-public interface StatusRepository extends CrudRepository<OrderStatus, Long>{
+public interface StatusRepository extends CrudRepository<OrderStatus, Integer>{
 }
