+++ spring-data/src/main/java/by/pvt/repository/TransmissionTypeRepository.java
@@ -3,5 +3,5 @@ package by.pvt.repository;
 import by.pvt.pojo.TransmissionType;
 import org.springframework.data.repository.CrudRepository;
 
-public interface TransmissionTypeRepository extends CrudRepository<TransmissionType, Long> {
+public interface TransmissionTypeRepository extends CrudRepository<TransmissionType, Integer> {
 }
